Updated Opal Kelly references and files to use the XEM7001
Changed the ToolsVersion to 15.0 instead of 4.0 for compatibility between different .NET Frameworks.

It is necessary to have a .NET Framework of 4.8 and above for compatibility with the .dll filse for the Opal Kelly XEM7001.

I created a new folder called libs that contains the necessary .dll files for the Opal Kelly XEM7001, but only for x86. WIthin libs, there are other necessary .dll files that get copied directly into the project when Atticus is compiled.

Throughout the solution, it is necessary to change the namespace from com.opalkelly.frontpanel to the correct namespace OpalKelly.FronPanel for the XEM7001.